refactor(rngconverter): move Driver to com.sun.msv.rngconverter; add module-info
- Avoid JPMS split package with msv-core (com.sun.msv.writer.relaxng)
- Add net.java.dev.msv.rngconverter module descriptor
- Relocate message bundles and tests to the new package
- Update assembly mainClass and RELAXNGConverter entrypoint
