+++ rngconverter/pom.xml
@@ -63,7 +63,7 @@ EVEN IF SUN HAS BEEN ADVISED OF THE POSSIBILITY OF SUCH DAMAGES.
                         <configuration>
                             <transformers>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>com.sun.msv.writer.relaxng.Driver</mainClass>
+                                    <mainClass>com.sun.msv.rngconverter.Driver</mainClass>
                                 </transformer>
                             </transformers>
                         </configuration>
@@ -73,6 +73,10 @@ EVEN IF SUN HAS BEEN ADVISED OF THE POSSIBILITY OF SUCH DAMAGES.
         </plugins>
     </build> -->
     <dependencies>
+        <dependency>
+            <groupId>xerces</groupId>
+            <artifactId>xercesImpl</artifactId>
+        </dependency>
         <dependency>
             <groupId>net.java.dev.msv</groupId>
             <artifactId>xsdlib</artifactId>
@@ -231,7 +235,7 @@ EVEN IF SUN HAS BEEN ADVISED OF THE POSSIBILITY OF SUCH DAMAGES.
                     <archive>
                         <index>true</index>
                         <manifest>
-                            <mainClass>com.sun.msv.writer.relaxng.Driver</mainClass>
+                            <mainClass>com.sun.msv.rngconverter.Driver</mainClass>
                         </manifest>
                         <manifestEntries>
                             <version>${project.version}</version>
