+++ rngconverter/src/main/java/module-info.java
@@ -1,0 +1,8 @@
+module net.java.dev.msv.rngconverter {
+    requires java.xml;
+    requires net.java.dev.msv.core;
+    requires net.java.dev.msv.xsdlib;
+    requires xercesImpl;
+
+    exports com.sun.msv.rngconverter;
+}
