+++ rngconverter/src/main/resources/com/sun/msv/rngconverter/Messages_ja.properties
