+++ rngconverter/src/test/java/RELAXNGConverter.java
@@ -1,5 +1,5 @@
 public class RELAXNGConverter {
 	public static void main( String[] args ) throws Exception {
-		com.sun.msv.writer.relaxng.Driver.main(args);
+		com.sun.msv.rngconverter.Driver.main(args);
 	}
 }
